moving model out of comonent file
