+++ src/ui-kit/experimental/footer/footer.component.ts
@@ -1,4 +1,5 @@
 import { Component, Input } from '@angular/core';
+import { FooterModel } from './model/FooterModel';
 
 
 @Component({
@@ -13,55 +14,3 @@ export class SamFooterComponent {
    */
   @Input() model: FooterModel;
 }
-
-export interface FooterModel {
-  /**
-   * 
-   */
-  linkSections: FooterLinkSection[];
-  /**
-   * 
-   */
-  footerLogo: FooterLogo;
-}
-
-export interface FooterLinkSection {
-  /**
-   * 
-   */
-  text: string;
-  /**
-   * 
-   */
-  links: FooterLink[];
-}
-
-export interface FooterLink {
-  /**
-   * 
-   */
-  text: string;
-  /**
-   * 
-   */
-  route: string;
-}
-
-export interface FooterLogo {
-  /**
-   * 
-   */
-  text: string;
-
-  /**
-   * 
-   */
-  imageSourcePath: string;
-
-  /**
-   * 
-   */
-  imageAltText: string;
-}
-
-
